+++ examples/basic/main.tf
@@ -17,6 +17,17 @@ module "vpc_base" {
   lan_cidr = "${var.lan_access_cidr}"
 }
 
+## Generates instance user data from a template
+resource "template_file" "user_data" {
+  filename = "../templates/user_data.tpl"
+
+  vars {
+    hostname = "${var.stack_item_label}-example"
+    fqdn = "${var.stack_item_label}-nat.${var.domain_name}"
+    ssh_user = "{var.ssh_user}"
+  }
+}
+
 module "vpc_az" {
   # Example GitHub source
   #source = "github.com/unifio/terraform-aws-vpc?ref=master//az"
@@ -36,9 +47,7 @@ module "vpc_az" {
   instance_type = "${var.instance_type}"
   key_name = "${var.key_name}"
   nat_sg_id = "${module.vpc_base.nat_sg_id}"
-  user_data_template = "${var.user_data_template}"
-  domain = "${var.domain_name}"
-  ssh_user = "${var.ssh_user}"
   enable_nat_auto_recovery = "${var.enable_nat_auto_recovery}"
   enable_nat_eip = "${var.enable_nat_eip}"
+  user_data = "${template_file.user_data.rendered}"
 }