+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## Unreleased
 
 #### FEATURES:
-
 * Added support for assigning Elastic IP address to each NAT instance.
 * Added support for disabling the provisioning of NAT instances [GH-3]
 * Added support for auto-recovery of NAT instances
 * Added support for VPC flow logs [GH-1]
+
+#### IMPROVEMENTS:
+* Replaced user_data template and parameters with generic user_data param.
 
 ## 0.1.0 (Oct 21, 2015)